Move page number method to path utils
